Update application to reflect the closure of Aula-Finder and introduce SunsetPlaceholder component
- Changed metadata titles and descriptions across various pages to announce the transition to a new platform.
- Replaced existing page content with a SunsetPlaceholder component that provides information about the closure and directs users to the new platform.

This update marks the end of the Aula-Finder service, guiding users to the new institutional platform for room availability at Universidad de los Andes via Bookeau Resources.
