+++ app/map/page.tsx
@@ -1,22 +1,22 @@
 import type { Metadata } from "next";
-import MapClient from "./MapClient";
+import SunsetPlaceholder from "@/components/SunsetPlaceholder";
 
 export const metadata: Metadata = {
-  title: "Mapa del Campus - Aula-Finder",
-  description: "Visualiza los edificios de la Universidad de los Andes en el mapa del campus y consulta la disponibilidad de salones.",
+  title: "Aula-Finder se está despidiendo",
+  description: "Aula-Finder se está despidiendo y ahora dirige a la plataforma oficial para consultar salones disponibles en la Universidad de los Andes.",
   openGraph: {
-    title: "Mapa del Campus - Aula-Finder",
-    description: "Visualiza los edificios de la Universidad de los Andes en el mapa del campus.",
+    title: "Aula-Finder se está despidiendo",
+    description: "La consulta de salones disponibles ahora continúa en la plataforma oficial de la Universidad de los Andes.",
     url: "/Aula-Finder/map",
     type: "website",
   },
   twitter: {
     card: "summary",
-    title: "Mapa del Campus - Aula-Finder",
-    description: "Visualiza los edificios de la Universidad de los Andes en el mapa del campus.",
+    title: "Aula-Finder se está despidiendo",
+    description: "La consulta de salones disponibles ahora continúa en la plataforma oficial de la Universidad de los Andes.",
   },
 };
 
 export default function MapPage() {
-  return <MapClient />;
+  return <SunsetPlaceholder />;
 }
