+++ lib/sunset-content.ts
@@ -1,0 +1,10 @@
+export const SUNSET_CONTENT = {
+  badge: "Nos despedimos",
+  title: "Aula-Finder se está despidiendo",
+  message:
+    "Aula-Finder está entrando en su etapa de cierre. La consulta oficial de salones disponibles para estudiantes ahora vive en una nueva plataforma institucional de la Universidad de los Andes.",
+  supportingText:
+    "Gracias por usar Aula-Finder. A partir de ahora, puedes revisar los espacios disponibles desde la herramienta.",
+  ctaLabel: "Ir a la nueva plataforma",
+  ctaHref: "https://serviciosalaulayauditorios.bookeau.com",
+} as const;
